+++ modules/event-subscription/variables.tf
@@ -79,14 +79,14 @@ variable "storage_queue_endpoint" {
 variable "webhook_endpoint" {
   description = "Webhook configuration block where the Event Subscription will receive events."
   type = object({
-    url                               = optional(string, "")
+    url                               = string
     base_url                          = optional(string)
     max_events_per_batch              = optional(number)
     preferred_batch_size_in_kilobytes = optional(number)
     active_directory_tenant_id        = optional(string)
     active_directory_app_id_or_uri    = optional(string)
   })
-  default = {}
+  default = null
 }
 
 variable "included_event_types" {